+++ packages/components/tests/unit/helpers/validateCssAndMatchSnapshot.ts
@@ -4,7 +4,7 @@ import { getCssObject } from '../../../src/test-utils';
 import type { TagName } from '@porsche-design-system/shared';
 
 export const validateCssAndMatchSnapshot = (css: string) => {
-  const cssObject = getCssObject(css);
+  const cssObject: any = getCssObject(css);
   const componentName = expect.getState().testPath.match(/\/([^\/]+)\/[^\/]+\.spec\.ts/)[1];
   const componentTagName = `p-${componentName}` as TagName;
   // Extract componentMeta from testPath, if it's a functional component this will be undefined
@@ -13,14 +13,14 @@ export const validateCssAndMatchSnapshot = (css: string) => {
   expect(css).not.toMatch('. {'); // Invalid css which was produced before
 
   validateVisibilityStyle(cssObject);
-  // TODO: fixme
-  // validateSlottedStyles(cssObject, componentTagName);
+  validateSlottedStyles(cssObject, componentTagName);
   validateHoverMediaQuery(cssObject);
 
   // Validations for components only
   if (componentMeta && !componentMeta.isInternal) {
     expect(cssObject[':host([hidden])']).toEqual({ display: 'none !important' });
     validateHostDisplayStyle(cssObject);
+    validateFormComponentHostDisplayStyle(cssObject, componentTagName);
   }
 
   expect(css).toMatchSnapshot();
@@ -36,12 +36,30 @@ const validateVisibilityStyle = (cssObject: object) => {
 };
 
 // Expect no !important rule on display style of :host selector since it should be overridable
-const validateHostDisplayStyle = (cssObject: object) => {
+const validateHostDisplayStyle = (cssObject: any) => {
   if (cssObject[':host'].display) {
     expect(cssObject[':host'].display).not.toMatch(/!important/);
   } else {
     // some components don't have a display style
     expect(cssObject[':host'].display).toBeUndefined();
+  }
+};
+
+// Expect all form components to have display: block as host style
+const validateFormComponentHostDisplayStyle = (cssObject: any, tagName: TagName) => {
+  if (
+    [
+      'p-checkbox-wrapper',
+      'p-multi-select',
+      'p-pin-code',
+      'p-radio-button-wrapper',
+      'p-select',
+      'p-select-wrapper',
+      'p-textarea',
+      'p-text-field-wrapper',
+    ].includes(tagName)
+  ) {
+    expect(cssObject[':host'].display).toBe('block');
   }
 };
 
@@ -51,7 +69,10 @@ const validateSlottedStyles = (cssObject: object, tagName: TagName) => {
   recursivelyApplyForKeyIncludes(cssObject, '::slotted', (_, value) => {
     Object.entries(value).forEach(([cssProp, cssValue]) => {
       // exceptions for tagName and css property are defined here
-      if (tagName !== 'p-textarea-wrapper' || !['height', 'min-height', 'resize'].includes(cssProp)) {
+      if (
+        !['p-textarea-wrapper', 'p-optgroup'].includes(tagName) &&
+        !['height', 'min-height', 'resize', 'margin'].includes(cssProp)
+      ) {
         expect(cssValue).toMatch(/!important$/);
       }